[Step3] feature: 테마 삭제 시, 현재 일시 이후로 남은 예약이 있으면 테마 삭제에 실패하도록 구현

+++ src/main/java/nextstep/theme/ThemeService.java
@@ -1,5 +1,9 @@
 package nextstep.theme;
 
+import nextstep.CurrentDateTimeService;
+import nextstep.reservation.ReservationRepository;
+import nextstep.schedule.Schedule;
+import nextstep.schedule.ScheduleJdbcRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -7,11 +11,21 @@ import java.util.List;
 @Service
 public class ThemeService {
     private final ThemeJdbcRepository themeJdbcRepository;
+    private final ScheduleJdbcRepository scheduleJdbcRepository;
+    private final ReservationRepository reservationRepository;
+    private final CurrentDateTimeService currentDateTimeService;
 
-    public ThemeService(ThemeJdbcRepository themeJdbcRepository) {
+    public ThemeService(
+            ThemeJdbcRepository themeJdbcRepository,
+            ScheduleJdbcRepository scheduleJdbcRepository,
+            ReservationRepository reservationRepository,
+            CurrentDateTimeService currentDateTimeService
+    ) {
         this.themeJdbcRepository = themeJdbcRepository;
+        this.scheduleJdbcRepository = scheduleJdbcRepository;
+        this.reservationRepository = reservationRepository;
+        this.currentDateTimeService = currentDateTimeService;
     }
-
 
     public Theme createTheme(ThemeCreateRequest request) {
         if (this.themeJdbcRepository.existByName(request.getName())) {
@@ -27,9 +41,30 @@ public class ThemeService {
     }
 
     public void deleteTheme(Long themeId) {
+        if (existReservationInTheme(themeId)) {
+            throw new IllegalStateException("테마 삭제 실패: 테마에 예약이 존재합니다.");
+        }
         boolean result = this.themeJdbcRepository.deleteThemeById(themeId);
         if (!result) {
-            throw new IllegalStateException("테마 삭제에 실패했습니다.: 해당 ID의 테마가 존재하지 않음");
+            throw new IllegalStateException("테마 삭제 실패.: 해당 ID의 테마가 존재하지 않음");
         }
+    }
+
+    /**
+     * 해당 테마에 예약이 존재하는지 여부를 반환합니다.
+     *
+     * @param themeId 조회하는 테마 ID
+     * @return true: 테마에 예약이 존재함 / false: 테마에 예약이 존재하지 않음
+     */
+    private boolean existReservationInTheme(Long themeId) {
+        List<Schedule> schedules = scheduleJdbcRepository.findByThemeIdAndDateGreaterThanAndTimeGreaterThan(
+                themeId,
+                currentDateTimeService.nowDate(),
+                currentDateTimeService.nowTime()
+        );
+        List<Long> scheduleIds = schedules.stream()
+                .map(Schedule::getId)
+                .toList();
+        return reservationRepository.existsByScheduleIds(scheduleIds);
     }
 }
